My_First_FASTAPI/requirements.txt
fastapi
uvicorn[standard]
requests
azure-identity
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-resource
boto3
python-dotenv
azure-keyvault-secrets
